+++ warp/_src/context.py
@@ -4272,6 +4272,7 @@ class Runtime:
                 ctypes.c_bool,  # fuse_fp
                 ctypes.c_bool,  # lineinfo
                 ctypes.c_bool,  # compile_time_trace
+                ctypes.c_bool,  # precompiled_headers
                 ctypes.c_char_p,  # output_path
                 ctypes.c_size_t,  # num_ltoirs
                 ctypes.POINTER(ctypes.c_char_p),  # ltoirs